+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.2.4] - 2021-02-11
+### Added
+- The ```FamilyClient``` now has an optional argument named ```$parameters```, in all its methods, to allow adding optional parameters to the URL
+- The ```CustomersClient``` has the same optional argument
+- The ```CheckDomainClient``` has the same optional argument
+
 ### Fixed
 - The ```License``` entity now allows ```null``` for the fields ```price.currency```, ```friendlyName```, and ```vendorSubscriptionId```